sql-json-schema-mapper tests are merged
